ref: css-mediaquery module refactoring
