+++ packages/core/ui/styling/css-selector.ts
@@ -1,11 +1,10 @@
 import { parse as convertToCSSWhatSelector, Selector as CSSWhatSelector, DataType as CSSWhatDataType } from 'css-what';
 import '../../globals';
 import { isCssVariable } from '../core/properties';
-import { Trace, CoreTypes } from './styling-shared';
 import { isNullOrUndefined } from '../../utils/types';
 
 import * as ReworkCSS from '../../css';
-import { checkIfMediaQueryMatches } from '../../media-query-list';
+import { checkIfMediaQueryMatches } from '../../css-mediaquery';
 
 export const MEDIA_QUERY_SEPARATOR = '&&';
 
@@ -675,7 +674,9 @@ export namespace Selector {
 		public trackChanges(node: Node, map: ChangeAccumulator): void {
 			this.selectors.forEach((sel, i) => {
 				if (i === 0) {
-					node && sel.trackChanges(node, map);
+					if (node) {
+						sel.trackChanges(node, map);
+					}
 				} else {
 					node = node.parent;
